moved ETLUtils and StructuredRecordStringConverter to new module, cdap-etl-common
