+++ cdap-app-templates/cdap-etl/cdap-etl-common/src/main/java/co/cask/cdap/etl/common/StructuredRecordStringConverter.java
@@ -1,4 +1,4 @@
-/*
+package co.cask.cdap.etl.common;/*
  * Copyright © 2015 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
@@ -13,8 +13,6 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-
-package co.cask.cdap.template.etl.common;
 
 import co.cask.cdap.api.common.Bytes;
 import co.cask.cdap.api.data.format.StructuredRecord;
