+++ cdap-app-templates/cdap-etl/pom.xml
@@ -36,6 +36,7 @@
     <module>cdap-etl-lib</module>
     <module>cdap-etl-realtime</module>
     <module>cdap-etl-archetypes</module>
+    <module>cdap-etl-common</module>
   </modules>
 
   <properties>
